#80 [REFACTOR] style 파일 분리, 컴포넌트명 변경 등을 진행
