+++ components/sns/common/SidebarTitle.tsx
@@ -1,7 +1,7 @@
-import styled from '@emotion/styled';
 import Link from 'next/link';
 
-import { WrapProps } from 'typings/wrapperProps';
+import type { WrapProps } from 'typings/wrapperProps';
+import * as S from './SidebarTitle.styles';
 
 interface SidebarTitleProps extends WrapProps {
   href: string;
@@ -13,21 +13,6 @@ const SidebarTitle = ({ children, href }: SidebarTitleProps) => {
       <Link href={href}>{children}</Link>
     </S.SidebarTitle>
   );
-};
-
-const S = {
-  SidebarTitle: styled.h1`
-    margin-bottom: 24px;
-
-    & > a {
-      text-decoration: none;
-      font-weight: 700;
-      font-size: 22px;
-      cursor: pointer;
-
-      color: ${({ theme }) => theme.colors.black};
-    }
-  `,
 };
 
 export default SidebarTitle;
