+++ components/sns/common/SocialLayout.styles.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+const SnsLayout = styled.div``;
+const Link = styled.a``;
+
+export { SnsLayout, Link };
