+++ components/sns/common/SocialMenus.styles.ts
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+const Menus = styled.ul``;
+
+export { Menus };
