+++ components/sns/common/SocialTitleSection.styles.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const SocialTitleSection = styled.div``;
+const SocialContentTitle = styled.h1`
+  margin-bottom: 40px;
+  font-weight: 700;
+  font-size: 36px;
+`;
+
+export { SocialTitleSection, SocialContentTitle };
