+++ pages/social/profile/index.tsx
@@ -1,15 +1,13 @@
-import SnsLayout from 'components/sns/common/SnsLayout';
+import SocialLayout from 'components/sns/common/SocialLayout';
 import SocialTitleSection from 'components/sns/common/SocialTitleSection';
 
 const SocialProfilePage = () => {
   return (
-    <SnsLayout>
+    <SocialLayout>
       <SocialTitleSection title="title 입력">
         <div>내용물 넣어주면 됨</div>
-        <div>내용물 넣어주면 됨</div>
-        <div>내용물 넣어주면 됨</div>
       </SocialTitleSection>
-    </SnsLayout>
+    </SocialLayout>
   );
 };
